🔎 ref: exception handler in update and delete

+++ src/main/java/br/com/cleilsonandrade/dscommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package br.com.cleilsonandrade.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import br.com.cleilsonandrade.dscommerce.dto.ProductDTO;
 import br.com.cleilsonandrade.dscommerce.entities.Product;
 import br.com.cleilsonandrade.dscommerce.repositories.ProductRepository;
+import br.com.cleilsonandrade.dscommerce.services.exceptions.DatabaseException;
 import br.com.cleilsonandrade.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -41,18 +46,29 @@ public class ProductService {
 
   @Transactional
   public ProductDTO update(Long id, ProductDTO dto) {
-    Product entity = repository.getReferenceById(id);
-
-    copyDtoToEntity(dto, entity);
-
-    entity = repository.save(entity);
-
-    return new ProductDTO(entity);
+    try {
+      Product entity = repository.getReferenceById(id);
+      copyDtoToEntity(dto, entity);
+      entity = repository.save(entity);
+      return new ProductDTO(entity);
+    } catch (EntityNotFoundException e) {
+      throw new ResourceNotFoundException("Resource not found");
+    }
   }
 
-  @Transactional
+  @Transactional(propagation = Propagation.SUPPORTS)
   public void delete(Long id) {
-    repository.deleteById(id);
+    if (!repository.existsById(id)) {
+      throw new ResourceNotFoundException("Resource not found");
+    }
+
+    try {
+      repository.deleteById(id);
+    } catch (EmptyResultDataAccessException e) {
+      throw new ResourceNotFoundException("Recurso nao encontrado");
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException("Referential integrity failure");
+    }
   }
 
   private void copyDtoToEntity(ProductDTO dto, Product entity) {
